airflow-docker/requirements.txt
apache-airflow
celery
redis
psycopg2-binary